using varint96 for groups count

+++ src/types/brc721_error.rs
@@ -19,7 +19,7 @@ pub enum Brc721Error {
     #[error("slot number too large for 96 bits: {0}")]
     InvalidSlotNumber(u128),
     #[error("group count must be at least 1, got {0}")]
-    InvalidGroupCount(u8),
+    InvalidGroupCount(u128),
     #[error("output index must be at least 1, got {0}")]
     InvalidOutputIndex(u8),
     #[error("range count must be at least 1, got {0}")]
